refactor: autocompletion code into a separate component
